## Purpose
Speeds up weight loading for NVFP4 TRTLLM MoE kernels by 10x

Before PR:
```
(Worker_TP2 pid=479) INFO 10-03 09:57:04 [default_loader.py:294] Loading weights took 44.17 seconds
(Worker_TP3 pid=480) WARNING 10-03 09:57:04 [weight_utils.py:1032] Found k_scale in the checkpoint (e.g. model.layers.5.self_attn.k_proj.k_scale), but not found the expected name in the model (e.g. model.layers.5.self_attn.attn.k_scale). k_scale is not loaded.
(Worker_TP0 pid=477) WARNING 10-03 09:57:04 [weight_utils.py:1032] Found k_scale in the checkpoint (e.g. model.layers.5.self_attn.k_proj.k_scale), but not found the expected name in the model (e.g. model.layers.5.self_attn.attn.k_scale). k_scale is not loaded.
(Worker_TP3 pid=480) WARNING 10-03 09:57:04 [weight_utils.py:1032] Found v_scale in the checkpoint (e.g. model.layers.5.self_attn.v_proj.v_scale), but not found the expected name in the model (e.g. model.layers.5.self_attn.attn.v_scale). v_scale is not loaded.
(Worker_TP0 pid=477) WARNING 10-03 09:57:04 [weight_utils.py:1032] Found v_scale in the checkpoint (e.g. model.layers.5.self_attn.v_proj.v_scale), but not found the expected name in the model (e.g. model.layers.5.self_attn.attn.v_scale). v_scale is not loaded.
Loading safetensors checkpoint shards:  99% Completed | 162/163 [00:44<00:00,  3.42it/s]
Loading safetensors checkpoint shards: 100% Completed | 163/163 [00:44<00:00,  3.55it/s]
Loading safetensors checkpoint shards: 100% Completed | 163/163 [00:44<00:00,  3.65it/s]
(Worker_TP0 pid=477) 
(Worker_TP3 pid=480) INFO 10-03 09:57:04 [default_loader.py:294] Loading weights took 44.75 seconds
(Worker_TP0 pid=477) INFO 10-03 09:57:04 [default_loader.py:294] Loading weights took 44.74 seconds

(Worker_TP3 pid=480) INFO 10-03 10:03:41 [gpu_model_runner.py:2758] Model loading took 93.9519 GiB and 441.511628 seconds                                                                                                                                                                                                                          
(Worker_TP1 pid=478) INFO 10-03 10:03:42 [gpu_model_runner.py:2758] Model loading took 93.9519 GiB and 442.370176 seconds                                                                                                                                                                                                                          
(Worker_TP2 pid=479) INFO 10-03 10:03:42 [gpu_model_runner.py:2758] Model loading took 93.9519 GiB and 442.423474 seconds                                                                                                                                                                                                                          
(Worker_TP0 pid=477) INFO 10-03 10:03:56 [gpu_model_runner.py:2758] Model loading took 93.9519 GiB and 457.105839 seconds    
```

After PR:
```
Worker_TP0 pid=37836) 
(Worker_TP3 pid=37839) INFO 10-02 16:16:47 [default_loader.py:294] Loading weights took 40.69 seconds
(Worker_TP0 pid=37836) INFO 10-02 16:16:47 [default_loader.py:294] Loading weights took 40.69 seconds
(Worker_TP0 pid=37836) INFO 10-02 16:16:47 [gpu_model_runner.py:2758] Model loading took 93.9513 GiB and 41.502974 seconds
(Worker_TP3 pid=37839) INFO 10-02 16:16:47 [gpu_model_runner.py:2758] Model loading took 93.9513 GiB and 41.497818 seconds
(Worker_TP1 pid=37837) INFO 10-02 16:16:50 [default_loader.py:294] Loading weights took 44.13 seconds
(Worker_TP2 pid=37838) INFO 10-02 16:16:50 [default_loader.py:294] Loading weights took 44.13 seconds
(Worker_TP1 pid=37837) INFO 10-02 16:16:51 [gpu_model_runner.py:2758] Model loading took 93.9513 GiB and 44.943625 seconds
(Worker_TP2 pid=37838) INFO 10-02 16:16:51 [gpu_model_runner.py:2758] Model loading took 93.9513 GiB and 44.947515 seconds
```


## Test Plan

lm_eval with GSM8k
```
# Server:
VLLM_USE_FLASHINFER_MOE_FP4=1 VLLM_FLASHINFER_MOE_BACKEND="latency" vllm serve nvidia/DeepSeek-R1-0528-FP4 --tensor-parallel-size 4

# Client:
python3 tests/evals/gsm8k/gsm8k_eval.py --port 8000

```

## Test Result

```

Results:
Accuracy: 0.953
Invalid responses: 0.000
Total latency: 79.459 s
Questions per second: 16.600
root@umbriel-b200-034:/workspace/scratch-pmaj-1/gh-pm-vllm# python3 tests/evals/gsm8k/gsm8k_eval.py --port 8000
Running GSM8K evaluation: 1319 questions, 5-shot
Evaluating: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 1319/1319 [01:52<00:00, 11.73it/s]

Results:
Accuracy: 0.952
Invalid responses: 0.000
Total latency: 112.475 s
Questions per second: 11.727
root@umbriel-b200-034:/workspace/scratch-pmaj-1/gh-pm-vllm# python3 tests/evals/gsm8k/gsm8k_eval.py --port 8000
Running GSM8K evaluation: 1319 questions, 5-shot
Evaluating: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 1319/1319 [01:52<00:00, 11.74it/s]

Results:
Accuracy: 0.961
Invalid responses: 0.000
Total latency: 112.386 s
Questions per second: 11.736
root@umbriel-b200-034:/workspace/scratch-pmaj-1/gh-pm-vllm# python3 tests/evals/gsm8k/gsm8k_eval.py --port 8000
Running GSM8K evaluation: 1319 questions, 5-shot
Evaluating: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 1319/1319 [01:51<00:00, 11.88it/s]

Results:
Accuracy: 0.961
Invalid responses: 0.000
Total latency: 111.027 s
Questions per second: 11.880
```

---
<details>
<summary> Essential Elements of an Effective PR Description Checklist </summary>

- [x] The purpose of the PR, such as "Fix some issue (link existing issues this PR will resolve)".
- [x] The test plan, such as providing test command.
- [x] The test results, such as pasting the results comparison before and after, or e2e results
- [N/A] (Optional) The necessary documentation update, such as updating `supported_models.md` and `examples` for a new model.
- [N/A] (Optional) Release notes update. If your change is user facing, please update the release notes draft in the [Google Doc](https://docs.google.com/document/d/1YyVqrgX4gHTtrstbq8oWUImOyPCKSGnJ7xtTpmXzlRs/edit?tab=t.0).
</details>